+++ core/hoyo/expedition.js
@@ -110,12 +110,14 @@ export default class Expedition extends HoyoTemplate {
 				continue;
 			}
 
+			if (account.skipChecks) {
+				continue;
+			}
+
 			Expedition.data.set(account.uid, {
 				...account,
 				fired: false
 			});
-
-			sr.Logger.info(`Loaded account with uid "${account.uid}" for expedition check`);
 		}
 	}